clean up main.dart
