+++ lib/src/models/generate_state.dart
@@ -25,26 +25,3 @@ class GenerateState with _$GenerateState {
     Language? language,
   }) = GenerationStateSuccess;
 }
-
-@freezed
-class CardsListView with _$CardsListView {
-  const CardsListView._();
-
-  const factory CardsListView({
-    required List<AnkiCard> cards,
-    required int currentPage,
-    required int totalPages,
-    required bool canPressNext,
-    required bool canPressPrevious,
-  }) = _CardsListView;
-
-  factory CardsListView.empty() {
-    return const CardsListView(
-      cards: [],
-      currentPage: 1,
-      totalPages: 1,
-      canPressNext: false,
-      canPressPrevious: false,
-    );
-  }
-}
