+++ lib/src/models/generate_state.freezed.dart
@@ -813,223 +813,3 @@ abstract class GenerationStateSuccess implements GenerateState {
   _$$GenerationStateSuccessCopyWith<_$GenerationStateSuccess> get copyWith =>
       throw _privateConstructorUsedError;
 }
-
-/// @nodoc
-mixin _$CardsListView {
-  List<AnkiCard> get cards => throw _privateConstructorUsedError;
-  int get currentPage => throw _privateConstructorUsedError;
-  int get totalPages => throw _privateConstructorUsedError;
-  bool get canPressNext => throw _privateConstructorUsedError;
-  bool get canPressPrevious => throw _privateConstructorUsedError;
-
-  @JsonKey(ignore: true)
-  $CardsListViewCopyWith<CardsListView> get copyWith =>
-      throw _privateConstructorUsedError;
-}
-
-/// @nodoc
-abstract class $CardsListViewCopyWith<$Res> {
-  factory $CardsListViewCopyWith(
-          CardsListView value, $Res Function(CardsListView) then) =
-      _$CardsListViewCopyWithImpl<$Res, CardsListView>;
-  @useResult
-  $Res call(
-      {List<AnkiCard> cards,
-      int currentPage,
-      int totalPages,
-      bool canPressNext,
-      bool canPressPrevious});
-}
-
-/// @nodoc
-class _$CardsListViewCopyWithImpl<$Res, $Val extends CardsListView>
-    implements $CardsListViewCopyWith<$Res> {
-  _$CardsListViewCopyWithImpl(this._value, this._then);
-
-  // ignore: unused_field
-  final $Val _value;
-  // ignore: unused_field
-  final $Res Function($Val) _then;
-
-  @pragma('vm:prefer-inline')
-  @override
-  $Res call({
-    Object? cards = null,
-    Object? currentPage = null,
-    Object? totalPages = null,
-    Object? canPressNext = null,
-    Object? canPressPrevious = null,
-  }) {
-    return _then(_value.copyWith(
-      cards: null == cards
-          ? _value.cards
-          : cards // ignore: cast_nullable_to_non_nullable
-              as List<AnkiCard>,
-      currentPage: null == currentPage
-          ? _value.currentPage
-          : currentPage // ignore: cast_nullable_to_non_nullable
-              as int,
-      totalPages: null == totalPages
-          ? _value.totalPages
-          : totalPages // ignore: cast_nullable_to_non_nullable
-              as int,
-      canPressNext: null == canPressNext
-          ? _value.canPressNext
-          : canPressNext // ignore: cast_nullable_to_non_nullable
-              as bool,
-      canPressPrevious: null == canPressPrevious
-          ? _value.canPressPrevious
-          : canPressPrevious // ignore: cast_nullable_to_non_nullable
-              as bool,
-    ) as $Val);
-  }
-}
-
-/// @nodoc
-abstract class _$$_CardsListViewCopyWith<$Res>
-    implements $CardsListViewCopyWith<$Res> {
-  factory _$$_CardsListViewCopyWith(
-          _$_CardsListView value, $Res Function(_$_CardsListView) then) =
-      __$$_CardsListViewCopyWithImpl<$Res>;
-  @override
-  @useResult
-  $Res call(
-      {List<AnkiCard> cards,
-      int currentPage,
-      int totalPages,
-      bool canPressNext,
-      bool canPressPrevious});
-}
-
-/// @nodoc
-class __$$_CardsListViewCopyWithImpl<$Res>
-    extends _$CardsListViewCopyWithImpl<$Res, _$_CardsListView>
-    implements _$$_CardsListViewCopyWith<$Res> {
-  __$$_CardsListViewCopyWithImpl(
-      _$_CardsListView _value, $Res Function(_$_CardsListView) _then)
-      : super(_value, _then);
-
-  @pragma('vm:prefer-inline')
-  @override
-  $Res call({
-    Object? cards = null,
-    Object? currentPage = null,
-    Object? totalPages = null,
-    Object? canPressNext = null,
-    Object? canPressPrevious = null,
-  }) {
-    return _then(_$_CardsListView(
-      cards: null == cards
-          ? _value._cards
-          : cards // ignore: cast_nullable_to_non_nullable
-              as List<AnkiCard>,
-      currentPage: null == currentPage
-          ? _value.currentPage
-          : currentPage // ignore: cast_nullable_to_non_nullable
-              as int,
-      totalPages: null == totalPages
-          ? _value.totalPages
-          : totalPages // ignore: cast_nullable_to_non_nullable
-              as int,
-      canPressNext: null == canPressNext
-          ? _value.canPressNext
-          : canPressNext // ignore: cast_nullable_to_non_nullable
-              as bool,
-      canPressPrevious: null == canPressPrevious
-          ? _value.canPressPrevious
-          : canPressPrevious // ignore: cast_nullable_to_non_nullable
-              as bool,
-    ));
-  }
-}
-
-/// @nodoc
-
-class _$_CardsListView extends _CardsListView {
-  const _$_CardsListView(
-      {required final List<AnkiCard> cards,
-      required this.currentPage,
-      required this.totalPages,
-      required this.canPressNext,
-      required this.canPressPrevious})
-      : _cards = cards,
-        super._();
-
-  final List<AnkiCard> _cards;
-  @override
-  List<AnkiCard> get cards {
-    if (_cards is EqualUnmodifiableListView) return _cards;
-    // ignore: implicit_dynamic_type
-    return EqualUnmodifiableListView(_cards);
-  }
-
-  @override
-  final int currentPage;
-  @override
-  final int totalPages;
-  @override
-  final bool canPressNext;
-  @override
-  final bool canPressPrevious;
-
-  @override
-  String toString() {
-    return 'CardsListView(cards: $cards, currentPage: $currentPage, totalPages: $totalPages, canPressNext: $canPressNext, canPressPrevious: $canPressPrevious)';
-  }
-
-  @override
-  bool operator ==(dynamic other) {
-    return identical(this, other) ||
-        (other.runtimeType == runtimeType &&
-            other is _$_CardsListView &&
-            const DeepCollectionEquality().equals(other._cards, _cards) &&
-            (identical(other.currentPage, currentPage) ||
-                other.currentPage == currentPage) &&
-            (identical(other.totalPages, totalPages) ||
-                other.totalPages == totalPages) &&
-            (identical(other.canPressNext, canPressNext) ||
-                other.canPressNext == canPressNext) &&
-            (identical(other.canPressPrevious, canPressPrevious) ||
-                other.canPressPrevious == canPressPrevious));
-  }
-
-  @override
-  int get hashCode => Object.hash(
-      runtimeType,
-      const DeepCollectionEquality().hash(_cards),
-      currentPage,
-      totalPages,
-      canPressNext,
-      canPressPrevious);
-
-  @JsonKey(ignore: true)
-  @override
-  @pragma('vm:prefer-inline')
-  _$$_CardsListViewCopyWith<_$_CardsListView> get copyWith =>
-      __$$_CardsListViewCopyWithImpl<_$_CardsListView>(this, _$identity);
-}
-
-abstract class _CardsListView extends CardsListView {
-  const factory _CardsListView(
-      {required final List<AnkiCard> cards,
-      required final int currentPage,
-      required final int totalPages,
-      required final bool canPressNext,
-      required final bool canPressPrevious}) = _$_CardsListView;
-  const _CardsListView._() : super._();
-
-  @override
-  List<AnkiCard> get cards;
-  @override
-  int get currentPage;
-  @override
-  int get totalPages;
-  @override
-  bool get canPressNext;
-  @override
-  bool get canPressPrevious;
-  @override
-  @JsonKey(ignore: true)
-  _$$_CardsListViewCopyWith<_$_CardsListView> get copyWith =>
-      throw _privateConstructorUsedError;
-}
