+++ lib/src/providers/session_id_provider.dart
@@ -1,0 +1,9 @@
+import 'package:ankigpt/src/models/session_id.dart';
+import 'package:riverpod_annotation/riverpod_annotation.dart';
+
+part 'session_id_provider.g.dart';
+
+@riverpod
+SessionId? sessionId(SessionIdRef ref) {
+  return null;
+}
